% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchRFA.R
\name{launchRFA}
\alias{launchRFA}
\title{Launch ForestAnalysisInR App}
\usage{
launchApp(...)
}
\arguments{
\item{...}{additional parameters from the \code{\link[shiny:runApp]{runApp}} function in the \emph{shiny} package.}
}
\value{
This function does not return.
}
\description{
This function launch the treetop application.
}
\details{
The ForestryAnalysisInR app doesn't really work actually. Big sad
}
\examples{
\dontrun{

# Launch treetop application
ForestryAnalysisInR::launchRFA(launch.browser = TRUE)

}
}
\references{

}
