% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfa.R
\name{rfa}
\alias{rfa}
\title{Forest Inventory Data}
\format{
\code{Package}Package name
\code{Applications}Application (e.g. remote sensing, mensuration, inventory)
\code{Description}CRAN/GitHub package description from author(S)
\code{Link}URL to access
}
\usage{
rfa()
}
\value{
A \code{data.frame}
}
\description{
Forest Inventory Data
}
\details{
Import forestry package list
}
\note{
List updated 2021-08-03
}
\examples{
rfa()
}
